Added plugin support! Albeit with a deprecated standard library feature.
Python v3.3 supported the necessary operation (loading an arbitrary .py
file) but that bit has been deprecated as well. To my knowledge, there
is no clean or simple way to do this with Python v3.4 without a lot of
string handling or messing with sys.path extensively, which I would
rather not do. We'll see how it turns out...