Cleaned up the Content Provider interface.
It's now slightly simpler to use. There is no more input -> source
abstraction model, and the base doesn't touch or store the inputs at
all.

In addition, the copy operations in the Environment class are now in
their own separate class called Operations. An instance is required by
the BaseContentProvider constructor, but it's optional. The default
being a created instance of that class.